refactor toshi_api_subtask.py extracted from ths_import,py; add coverage on toshi_api_subtask (WIP);
