+++ tests/oq_import/conftest.py
@@ -3,6 +3,21 @@ import pathlib
 
 import pytest
 
+TASK_ID = "TID12345"
+GT_ID = "GTXYZ=="
+TASK_ARGS_JSON = "task_args.json"
+
+
+@pytest.fixture
+def task_id():
+    yield TASK_ID
+
+
+@pytest.fixture
+def general_task_id():
+    yield GT_ID
+
+
 mid_skool_task_args_example = {
     'vs30': 275,
     'oq': {
@@ -83,11 +98,6 @@ old_skool_task_args_example = {
 }
 
 
-@pytest.fixture(scope='module')
-def task_id():
-    yield 'ABCD'
-
-
 @pytest.fixture
 def old_skool_task_args(tmpdir_factory, task_id):
     config_folder = pathlib.Path(tmpdir_factory.mktemp("old_skool"))
@@ -114,3 +124,54 @@ def mid_skool_config(tmpdir_factory, task_id):
 def latest_config():
     fname = "R2VuZXJhbFRhc2s6NjkzMTg5Mg==/subtasks/T3BlbnF1YWtlSGF6YXJkVGFzazo2OTMxODkz/task_args.json"
     yield pathlib.Path(__file__).parent / "fixtures" / fname
+
+
+@pytest.fixture
+def solution_archive_fixture(tmpdir_factory):
+    storage_folder = pathlib.Path(tmpdir_factory.mktemp("solution_archive_fixture"))
+    fixture = (
+        pathlib.Path(__file__).parent / 'fixtures' / 'openquake_hdf5_archive-T3BlbnF1YWtlSGF6YXJkVGFzazoxMDYzMzU3.zip'
+    )
+
+    # copy fixture data
+    tmp_file = storage_folder / fixture.name
+    tmp_file.write_bytes(fixture.read_bytes())
+    return tmp_file
+
+
+@pytest.fixture(scope="session")
+def hdf5_calc_fixture():
+    hdf5_fixture = (
+        pathlib.Path(__file__).parent.parent
+        / 'fixtures/oq_import/openquake_hdf5_archive-T3BlbnF1YWtlSGF6YXJkVGFzazo2OTMxODkz/calc_1.hdf5'
+    )
+    yield hdf5_fixture
+
+
+@pytest.fixture
+def mock_task_args_file_path(tmp_path, latest_config):
+
+    # mock artefects
+    subtasks_folder = tmp_path / GT_ID / 'subtasks'
+    subtasks_folder.mkdir(parents=True)
+    # task_id = '12345'
+    ta_fixt = json.load(open(latest_config, 'r'))
+    ta = {
+        "hazard_model-hazard_config": ta_fixt.get("hazard_model-hazard_config"),
+        'site_params-vs30': 760,
+        'intensity_spec': {'measures': ['PGA'], 'levels': [0.01, 0.02]},
+        "hazard_curve-imts": ["PGA", "SA(0.5)", "SA(1.5)", "SA(3.0)"],
+        "hazard_curve-imtls": [
+            0.0001,
+            0.0002,
+            0.0004,
+        ],
+        "site_params-locations": ["WLG", "AKL", "DUD", "CHC"],
+        "site_params-locations_file": None,
+    }
+
+    task_args_file = subtasks_folder / str(TASK_ID) / TASK_ARGS_JSON
+    task_args_file.parent.mkdir()
+    task_args_file.write_text(json.dumps(ta))
+    # config = oq_config.config_from_task(task_id, subtasks_folder)
+    yield task_args_file
